Safari-609.1.11.1/LayoutTests/js/dom/document-all-strict-eq-expected.txt

Test to make sure that document.all works properly with strict eq in the DFG

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


document.all was undefined
PASS f(testObj, document.all) is false
PASS f(document.all, document.all) is true
PASS successfullyParsed is true

TEST COMPLETE
